+++ age/CHANGELOG.md
@@ -23,6 +23,9 @@ to 1.0.0 are beta releases.
   optionally apply the armored age format.
 
 ### Changed
+- `age::decryptor::RecipientsDecryptor` now takes
+  `impl Iterator<Item = Identity>` in its decryption methods instead of
+  `&[Identity]`.
 - `age::Encryptor::wrap_output` now only generates the non-malleable binary age
   format. Use `encryptor.wrap_output(ArmoredWriter::wrap_output(output, format))`
   to optionally generate armored age files.